third_party/blink/web_tests/platform/win11-arm64/virtual/web-bluetooth-new-permissions-backend/external/wpt/bluetooth/bidi/device/forget/getDevices.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] forget() removes devices from getDevices().
  assert_equals: getDevices() should return the granted device. expected 1 but got 2
Harness: the test ran to completion.
